Criação da página inicial do projeto index.ejs
