+++ public/js/index.js
@@ -1,0 +1,5 @@
+const hamburguer = document.querySelector(".hamburguer");
+hamburguer.onclick = function() {
+    const nav = document.querySelector("nav");
+    nav.classList.toggle("active")
+}
