+++ public/styles/header.css
@@ -1,0 +1,130 @@
+:root {
+    --primary-color: #161D6F;
+    --secondary-color: #98DED9;
+    --text-color: #F6F6F6;
+    --assets: #C7FFD8;
+}
+
+* {
+    list-style: none;
+    text-decoration: none;
+    box-sizing: border-box;
+}
+
+body {
+    margin: 0;
+}
+
+div#img-heading a {
+    color: var(--assets);
+}
+
+div.main {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px 100px;
+    background-color: var(--primary-color);
+    color: var(--assets);
+    margin: 0;
+}
+
+div#img-heading {
+    display: flex;
+    align-items: center;
+}
+
+img.logo {
+    width: 70px;
+    border-radius: 50%;
+    display: inline-block;
+}
+
+div h1 {
+    display: inline;
+    margin-left: 15px;
+    font-size: 20px;
+}
+
+.hamburguer {
+    display: none;
+}
+
+nav ul {
+    display: flex;
+}
+
+nav ul li a {
+    display: block;
+    color: var(--assets);
+    font-size: 20px;
+    padding: 10px 30px;
+    border-radius: 50px;
+    transition: .2s;
+    margin: 0 5px;
+}
+
+nav ul li a:hover {
+    background-color: var(--assets);
+    color: var(--primary-color);
+}
+
+nav ul li a.active {
+    background-color: var(--assets);
+    color: var(--primary-color);
+}
+
+@media only screen and (max-width: 1320px) {
+    div.main {
+        padding: 5px 50px;
+    }
+}
+
+@media only screen and (max-width: 1100px) {
+    div.main {
+        padding: 5px 30px;
+    }
+}
+
+@media only screen and (max-width: 900px) {
+    .hamburguer {
+        display: block;
+        cursor: pointer;
+    }
+    .hamburguer .line {
+        width: 30px;
+        height: 3px;
+        background-color: var(--assets);
+        margin: 6px 0;
+    }
+    nav {
+        height: 0px;
+        position: absolute;
+        top: 80px;
+        left: 0;
+        right: 0;
+        width: 100vw;
+        background-color: var(--primary-color);
+        transition: .2s;
+        overflow: hidden;
+    }
+
+    nav.active {
+        height: 300px;
+    }
+
+    nav ul {
+        display: block;
+        width: fit-content;
+        margin: 80px auto 0;
+        text-align: center;
+        transition: .5s;
+        opacity: 0;
+    }
+    nav.active ul {
+        opacity: 1;
+    }
+    nav ul li a {
+        margin-bottom: 12px;
+    }
+}
