+++ public/styles/index.css
@@ -6,204 +6,178 @@
 }
 
 body {
-    background: var(--secondary-color) url(../imagens/abstract-5035370_1920.jpg) fixed center / cover no-repeat;
-    backdrop-filter: blur(3px);
     font-family: Verdana, Geneva, Tahoma, sans-serif;
-    min-height: 100vh;
-    margin: 0;
-    display: flex;
-    align-items: center;
+    text-align: center;
 }
 
-.flex-wrapper {
+div.hero, div.features {
+    padding: 50px 120px;
     display: flex;
-    flex-direction: row;
-    align-items: center;
     justify-content: space-around;
-    height: 100%;
-    width: 90vw;
-    margin: auto;
-}
-
-#first-section {
-    width: 50%;
-    height: 100%;
-    display: flex;
     align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    color: var(--text-color);
 }
 
-#first-section img {
-    width: 70%;
+div.hero-left, div.call-to-action {
+    width: 45%;
+    text-align: left;
+}
+
+div.call-to-action {
+    margin-left: 30px;
+}
+
+div.hero-left h2, div.call-to-action h2 {
+    font-size: 35px;
+    margin-left: 0;
+    color: var(--primary-color);
+}
+
+div.hero-left p, div.call-to-action p {
+    font-size: 20px;
+    color: var(--primary-color);
+}
+
+div.hero-right {
+    width: 50%;
+}
+
+div.hero-right img.hero-image {
+    width: 100%;
     border-radius: 15px;
 }
 
-#first-section h1 {
-    display: inline;
-    font-size: 55px;
-    color: var(--secondary-color);
-    text-shadow: 2px 2px 6px rgba(0, 0, 0, 0.74);
+div.features {
+    background-color: var(--text-color);
+}
+
+div.feature-images {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-rows: 1fr 1fr;
+    gap: 10px;
+    width: 48%;
     text-align: center;
-    margin-top: 0;
 }
 
-#second-section {
-    height: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 50%;
-
-}
-
-#form-div {
-    background-color: var(--secondary-color);
-    margin: auto;
-    width: 350px;
-    height: 500px;
+div.feature-img {
+    background-color: white;
     border-radius: 10px;
-    overflow: hidden;
-    box-shadow: 5px 20px 50px #000;
-}
-
-#check {
-    display: none;
-}
-
-.signup {
-    position: relative;
+    margin: 10px;
     width: 100%;
     height: 100%;
+    padding: 20px 0 10px;
 }
 
-label {
+div.features h2, div.features p{
     color: var(--primary-color);
-    font-size: 2.3rem;
-    justify-content: center;
-    display: flex;
-    margin: 60px;
-    font-weight: bold;
-    cursor: pointer;
-    transition: .5s ease-in-out;
 }
 
-input {
-    width: 60%;
-    height: 20px;
-    background-color: #e0dede;
-    justify-content: center;
-    display: flex;
-    margin: 20px auto;
-    padding: 10px;
-    border: none;
-    outline: none;
-    border-radius: 5px;
+div.feature-img-wrapper {
+    background-color: var(--text-color);
+    border-radius: 50%;
+    padding: 15px;
+    width: 36%;
+    margin: auto;
 }
 
-button {
-    width: 60%;
-    height: 40px;
-    margin: 10px auto;
-    justify-content: center;
-    display: block;
+div.feature-images > div.feature-img img {
+    width: 85%;
+}
+
+div.feature-img p {
+    padding: 5px 6px;
+    margin-top: -10px;
+}
+
+div.call-to-action {
+    width: 50%;
+}
+
+div.call-to-action p {
+    margin-bottom: 50px;
+}
+
+div.call-to-action a {
     color: var(--assets);
     background-color: var(--primary-color);
-    font-size: 1em;
+    padding: 12px 20px;
+    border-radius: 10px;
+    transition: .3s;
     font-weight: bold;
-    margin-top: 20px;
-    outline: none;
-    border: none;
-    border-radius: 5px;
-    transition:  .2s ease-in;
-    cursor: pointer;
 }
 
-button:hover {
-    background-color: #2b38c2;
+div.call-to-action a:hover {
+    background-color: var(--assets);
+    color: var(--primary-color);
 }
 
-.login {
-    height: 460px;
-    background-color: var(--primary-color);
-    border-radius:  60% / 10%;
-    transform: translateY(-180px);
-    transition: .8s ease-in-out;
+@media screen and ( max-width: 1320px ) {
+    div.hero, div.features {
+        flex-direction: column;
+        padding: 10px 0;
+    }
+
+    div.call-to-action, div.hero-left {
+        width: 100%;
+        padding: 20px;
+        text-align: center;
+    }
+
+    div.hero-right {
+        width: 85%;
+    }
+
+    div.call-to-action {
+        margin: 0 0 10px;
+    }
+
+    div.feature-img {
+        width: 100%;
+    }
+    div.feature-images {
+        order: 1;
+    }
+    div.call-to-action{
+        order: 0;
+    }
+    div.call-to-action p{
+        margin-left: 0;
+    }
 }
 
-.login label {
-    color: var(--assets);
-    transform: scale(.6);
-}
-
-.login p {
-    text-align: center;
-    color: rgb(216, 2, 2);
-}
-
-#check:checked ~ .login {
-    transform: translateY(-500px);
-}
-
-#check:checked ~ .login label {
-    transform: scale(1);
-}
-
-#check:checked ~ .signup label {
-    transform: scale(.6);
-}
-
-@media screen and (max-width: 800px){
-    .flex-wrapper {
+@media screen and ( max-width: 1000px ) {
+    div.feature-images {
         display: flex;
         flex-direction: column;
+        width: 30%;
     }
-    #first-section h1 {
-        font-size: 35px;
-    }
-    #first-section {
+    div.feature-img {
         width: 100%;
     }
-    #form-div {
-
-        margin:20px auto;
+    div.hero-left h2, div.call-to-action h2 {
+        font-size: 1.5em;
     }
-    /* #second-section {
-        padding: 20px 0;
-        width: 100% ;
-        height: 650px;
+    div.hero-left p, div.call-to-action p {
+        font-size: 1.2em;
     }
-    #form-div {
-        height: 100%;
-        padding: 15px 0;
-    }
-    #form-div > p {
-        padding-left: 20px;
-        margin: 10px 0 0;
-    }
-    
-    p#signup {
-        font-size: 20px;
-        margin: 10px 0 0;
-    } */
 }
 
-/* @media screen and (min-width: 801px) and (max-width: 1500px){
-    #form-div {
+@media screen and (max-width: 850px){
+    div.hero-right img.hero-image {
+        width: 80%;
+    }
+    div.feature-images {
+        display: grid;
+        grid-template-columns: 180px 180px;
+        grid-template-rows: 180px 180px;
+        gap: 10px;
+        width: 370px;
+    }
+    div.feature-img {
         width: 100%;
-        min-height: 400px;
-        margin: 0;
     }
-    #first-section {
-        width: 49%;
+    div.feature-img {
+        font-size: 12px;
     }
-    #first-section h1{
-        font-size: 40px;
-        margin-top: 0;
-    }
-    #second-section {
-        width: 49%;
-        height: 600px;
-    }
-} */
+    
+}
