+++ views/contacts.ejs
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+
+    <link rel="stylesheet" href="styles/header.css">
+</head>
+<body>
+    <%- include('partials/header.ejs') %>
+
+
+</body>
+</html>
