+++ views/partials/header.ejs
@@ -1,0 +1,30 @@
+<header>
+    <div class="main">
+        <div id="img-heading">
+            <a href="/home">
+                <img src="imagens/logo.jpg" alt="logo" class="logo">
+            </a>
+            <a href="/home">
+                <h1> Col Lab </h1>
+            </a>
+        </div>
+        <div class="hamburguer">
+            <div class="line"></div>
+            <div class="line"></div>
+            <div class="line"></div>
+        </div>
+        <nav>
+            <ul>
+                <li>
+                    <a href="/home" class="active">Home</a>
+                </li>
+                <li>
+                    <a href="/about">Sobre nós</a>
+                </li>
+                <li>
+                    <a href="/contacts">Contatos</a>
+                </li>
+            </ul>
+        </nav>
+    </div>
+</header>
