+++ views/projects.ejs
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+
+    <link rel="stylesheet" href="styles/header.css">
+</head>
+<body>
+    <%- include('partials/header.ejs') %>
+
+    <table>
+        <tr>
+            <th>Projeto</th>
+            <th>Descrição</th>
+            <th>Data de criação</th>
+        </tr>
+        <tr>
+            <td>Simon game</td>
+            <td>Projeto de um jogo no navegador de memorização de padrões.</td>
+            <td>25/09/2024</td>
+        </tr>
+    </table>
+</body>
+</html>
+
